(restructure) flatten the Accordion, Error, and InformaionList folders and move from components/shared/ to components/
