+++ src/components/Accordion.js
@@ -1,8 +1,60 @@
-import { Children } from 'react'
+import { Children, cloneElement, useState } from 'react'
 import { Box, Text } from 'grommet'
 import isFunction from 'helpers/isFunction'
+import { CheckBox } from 'components/CheckBox'
 import { ExpandableBlock } from 'components/ExpandableBlock'
 import { Icon } from 'components/Icon'
+
+export const Accordion = (props) => {
+  const getChildrenCount = () => {
+    return Children.count(props.children)
+  }
+
+  const [activeElements, setActiveElements] = useState(
+    new Array(getChildrenCount()).fill(false)
+  )
+
+  const isAllExpanded = !activeElements.includes(false)
+
+  const handleToggleAll = () => {
+    if (isAllExpanded) {
+      setActiveElements(new Array(getChildrenCount()).fill(false))
+    } else {
+      setActiveElements(new Array(getChildrenCount()).fill(true))
+    }
+  }
+
+  const handleToggle = (index) => {
+    setActiveElements(() =>
+      activeElements.map((x, i) => (i === index ? !x : x))
+    )
+  }
+
+  const children = Children.map(
+    props.children,
+    (child, i) =>
+      child &&
+      cloneElement(child, {
+        expand: activeElements[i],
+        handleToggle: () => handleToggle(i)
+      })
+  )
+
+  return (
+    <Box margin={{ top: 'xsmall' }}>
+      {!props.hideExpandAll && children.length > 1 && (
+        <Box margin={{ bottom: 'small' }}>
+          <CheckBox
+            label="Expand all"
+            onClick={() => handleToggleAll()}
+            checked={isAllExpanded}
+          />
+        </Box>
+      )}
+      {children}
+    </Box>
+  )
+}
 
 export const AccordionPanel = ({
   title,
@@ -57,5 +109,3 @@ export const AccordionPanel = ({
     </Box>
   )
 }
-
-export default AccordionPanel
