+++ src/components/ExperimentDetail.js
@@ -13,10 +13,7 @@ import { ExperimentCardAction } from 'components/ExperimentCardAction'
 import { ExperimentCardHeader } from 'components/ExperimentCardHeader'
 import { ExperimentCardMeta } from 'components/ExperimentCardMeta'
 import { FixedContainer } from 'components/FixedContainer'
-import {
-  InformationList,
-  InformationItem
-} from 'components/shared/InformationList'
+import { InformationList, InformationItem } from 'components/InformationList'
 import { TextHighlight } from 'components/TextHighlight'
 
 export const ExperimentDetail = ({ experiment }) => {
