--- src/components/shared/Accordion/index.js
@@ -1,2 +1,0 @@
-export { Accordion } from './Accordion'
-export { AccordionPanel } from './AccordionPanel'
