--- src/components/shared/InformationList/index.js
@@ -1,2 +1,0 @@
-export { InformationItem } from './informationItem'
-export { InformationList } from './informationList'
