+++ src/components/shared/SamplesTable/ProcessingInformationCell/ModalContent/SubmitterSupplied/ArrayExpressSupplied.js
@@ -1,7 +1,7 @@
 import { Box, Heading, Text } from 'grommet'
 import formatString from 'helpers/formatString'
 import truncateOnWord from 'helpers/truncateOnWord'
-import { Accordion, AccordionPanel } from 'components/shared/Accordion'
+import { Accordion, AccordionPanel } from 'components/Accordion'
 import { Anchor } from 'components/Anchor'
 
 const Title =
