+++ src/components/shared/SamplesTable/ProcessingInformationCell/ModalContent/SubmitterSupplied/GeoSubmitterSupplied.js
@@ -1,9 +1,6 @@
 import { Box, Heading, Paragraph } from 'grommet'
 import { Anchor } from 'components/Anchor'
-import {
-  InformationList,
-  InformationItem
-} from 'components/shared/InformationList'
+import { InformationList, InformationItem } from 'components/InformationList'
 import { links } from 'config'
 
 export const GeoSubmitterSupplied = ({
