+++ src/pages/compendia/[type]/download/[organismName].js
@@ -6,7 +6,7 @@ import { useResponsive } from 'hooks/useResponsive'
 import { api } from 'api'
 import { compendia } from 'config'
 import { Column } from 'components/Column'
-import { Error } from 'components/shared/Error'
+import { Error } from 'components/Error'
 import { FixedContainer } from 'components/FixedContainer'
 import { Row } from 'components/Row'
 import { Explore } from 'components/Compendia/Explore'
