+++ src/pages/download.js
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router'
 import { useDatasetManager } from 'hooks/useDatasetManager'
 import { useResponsive } from 'hooks/useResponsive'
 import scrollTo from 'helpers/scrollTo'
-import { Error } from 'components/shared/Error'
+import { Error } from 'components/Error'
 import { FixedContainer } from 'components/FixedContainer'
 import { Row } from 'components/Row'
 import { Spinner } from 'components/Spinner'
